requirements.txt
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
streamlit
opencv-python-headless
Pillow
numpy